move camera-related files into camera/ directory
