+++ packages/flame/lib/camera.dart
@@ -1,0 +1,12 @@
+export 'src/camera/bounded_position_behavior.dart'
+    show BoundedPositionBehavior;
+export 'src/camera/camera_component.dart' show CameraComponent;
+export 'src/camera/circular_viewport.dart' show CircularViewport;
+export 'src/camera/fixed_aspect_ratio_viewport.dart'
+    show FixedAspectRatioViewport;
+export 'src/camera/fixed_size_viewport.dart' show FixedSizeViewport;
+export 'src/camera/follow_behavior.dart' show FollowBehavior;
+export 'src/camera/max_viewport.dart' show MaxViewport;
+export 'src/camera/viewfinder.dart' show Viewfinder;
+export 'src/camera/viewport.dart' show Viewport;
+export 'src/camera/world.dart' show World;
