+++ packages/flame/lib/experimental.dart
@@ -26,16 +26,16 @@ export 'src/events/messages/drag_update_event.dart' show DragUpdateEvent;
 export 'src/events/messages/tap_cancel_event.dart' show TapCancelEvent;
 export 'src/events/messages/tap_down_event.dart' show TapDownEvent;
 export 'src/events/messages/tap_up_event.dart' show TapUpEvent;
-export 'src/experimental/bounded_position_behavior.dart'
+export 'src/camera/bounded_position_behavior.dart'
     show BoundedPositionBehavior;
-export 'src/experimental/camera_component.dart' show CameraComponent;
-export 'src/experimental/circular_viewport.dart' show CircularViewport;
-export 'src/experimental/fixed_aspect_ratio_viewport.dart'
+export 'src/camera/camera_component.dart' show CameraComponent;
+export 'src/camera/circular_viewport.dart' show CircularViewport;
+export 'src/camera/fixed_aspect_ratio_viewport.dart'
     show FixedAspectRatioViewport;
 export 'src/experimental/fixed_integer_resolution_viewport.dart'
     show FixedIntegerResolutionViewport;
-export 'src/experimental/fixed_size_viewport.dart' show FixedSizeViewport;
-export 'src/experimental/follow_behavior.dart' show FollowBehavior;
+export 'src/camera/fixed_size_viewport.dart' show FixedSizeViewport;
+export 'src/camera/follow_behavior.dart' show FollowBehavior;
 export 'src/experimental/geometry/shapes/circle.dart' show Circle;
 export 'src/experimental/geometry/shapes/polygon.dart' show Polygon;
 export 'src/experimental/geometry/shapes/rectangle.dart' show Rectangle;
@@ -43,7 +43,7 @@ export 'src/experimental/geometry/shapes/rounded_rectangle.dart'
     show RoundedRectangle;
 export 'src/experimental/geometry/shapes/shape.dart' show Shape;
 export 'src/experimental/has_game_reference.dart' show HasGameReference;
-export 'src/experimental/max_viewport.dart' show MaxViewport;
-export 'src/experimental/viewfinder.dart' show Viewfinder;
-export 'src/experimental/viewport.dart' show Viewport;
-export 'src/experimental/world.dart' show World;
+export 'src/camera/max_viewport.dart' show MaxViewport;
+export 'src/camera/viewfinder.dart' show Viewfinder;
+export 'src/camera/viewport.dart' show Viewport;
+export 'src/camera/world.dart' show World;
