+++ packages/flame/lib/src/camera/bounded_position_behavior.dart
