+++ packages/flame/lib/src/camera/camera_component.dart
@@ -6,13 +6,13 @@ import 'package:flame/src/effects/controllers/effect_controller.dart';
 import 'package:flame/src/effects/move_effect.dart';
 import 'package:flame/src/effects/move_to_effect.dart';
 import 'package:flame/src/effects/provider_interfaces.dart';
-import 'package:flame/src/experimental/bounded_position_behavior.dart';
-import 'package:flame/src/experimental/follow_behavior.dart';
+import 'package:flame/src/camera/bounded_position_behavior.dart';
+import 'package:flame/src/camera/follow_behavior.dart';
 import 'package:flame/src/experimental/geometry/shapes/shape.dart';
-import 'package:flame/src/experimental/max_viewport.dart';
-import 'package:flame/src/experimental/viewfinder.dart';
-import 'package:flame/src/experimental/viewport.dart';
-import 'package:flame/src/experimental/world.dart';
+import 'package:flame/src/camera/max_viewport.dart';
+import 'package:flame/src/camera/viewfinder.dart';
+import 'package:flame/src/camera/viewport.dart';
+import 'package:flame/src/camera/world.dart';
 import 'package:meta/meta.dart';
 import 'package:vector_math/vector_math_64.dart';
