+++ packages/flame/lib/src/camera/fixed_size_viewport.dart
@@ -1,6 +1,6 @@
 import 'dart:ui';
 
-import 'package:flame/src/experimental/viewport.dart';
+import 'package:flame/src/camera/viewport.dart';
 import 'package:vector_math/vector_math_64.dart';
 
 /// A rectangular viewport with fixed dimensions.
